Add Tensorflow model and connection
