+++ README.md
@@ -1,2 +1,2 @@
 # NakijkBot
-Artificial Intelligence bot that predicts written numbers.
+Artificial Intelligence bot that predicts written numbers. This bot is created to assist teachers on primary schools.
