+++ package.json
@@ -8,10 +8,12 @@
     "chart.js": "^2.9.4"
   },
   "devDependencies": {
-    "eslint": "^7.15.0"
+    "eslint": "^7.15.0",
+    "parcel-bundler": "^1.12.4"
   },
   "scripts": {
-    "test": ""
+    "dev": "parcel index.html",
+    "build": "parcel build index.html"
   },
   "repository": {
     "type": "git",
