+++ style/app.css
@@ -1,0 +1,7 @@
+.prediction--box {
+  max-width: 20rem;
+}
+
+#chalkboard {
+  background-color: #0e4f29;
+}
